requirements.txt
streamlit
yfinance
plotly
pandas
numpy
python-dotenv
requests
watchdog